removed react-native dependency from navigation hooks
