+++ src/app/navigation/use-update-search-params.native.ts
@@ -1,7 +1,5 @@
 import { useCallback } from 'react'
-import { useRouter } from './use-router'
 import { useNavigation } from '../../router/use-navigation'
-import { Platform } from 'react-native'
 import { UseUpdateSearchParamsReturns } from './use-update-search-params.types'
 
 export default function <
